Move single headeer file in a package which can give correct dependency
