+++ Alignment/TrackerAlignment/plugins/TrackerSystematicMisalignments.cc
@@ -22,7 +22,7 @@
 #include "DataFormats/DetId/interface/DetId.h"
 #include "Geometry/Records/interface/TrackerDigiGeometryRecord.h"
 #include "Geometry/TrackerGeometryBuilder/interface/TrackerGeometry.h"
-#include "Geometry/TrackingGeometryAligner/interface/GeometryAligner.h"
+#include "Geometry/CommonTopologies/interface/GeometryAligner.h"
 #include "CLHEP/Random/RandGauss.h"
 
 #include "Alignment/TrackerAlignment/plugins/TrackerSystematicMisalignments.h"
