--- Geometry/TrackingGeometryAligner/BuildFile.xml
@@ -1,2 +1,0 @@
-<export>
-</export>
